RAGpipeline/requirements.txt
langchain 
langchain-community 
faiss-cpu
sentence-transformers
python-dotenv
langchain-groq
fastapi 
uvicorn[standard]
